requirements.txt
# Core
Django==5.1.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
python-dotenv==1.0.1

# DB
psycopg2-binary==2.9.9